\name{datanugget-package}
\alias{datanugget-package}

\title{
Data Nuggets
}

\description{
This package contains functions to create, refine, and cluster data nuggets which serve as representative samples of large datasets. The functions which perform these processes are create.DN, refine.DN, and WKmeans, respectively.
}

\author{
Traymon Beavers, Javier Cabrera, Mariusz Lubomirski
}

\references{
Data Nuggets: A Method for Reducing Big Data While Preserving Data Structure (Submitted for Publication, 2019)
}
